third_party/blink/web_tests/fast/filesystem/file-from-file-entry-expected.txt

Obtaining File from FileEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testFile.name is testFileEntry.name
PASS testFile.type is 'text/plain'
PASS testFile.size is 0
PASS successfullyParsed is true

TEST COMPLETE
